+++ ui/Modal/Search.svelte
@@ -17,7 +17,7 @@
   import { ValidationStatus } from "../src/validation";
 
   import { Icon, Input } from "../DesignSystem/Primitive";
-  import { FollowButton, Remote } from "../DesignSystem/Component";
+  import { FollowToggle, Remote } from "../DesignSystem/Component";
 
   let id: string;
   let value: string;
@@ -158,7 +158,7 @@
       <div slot="error" style="padding: 1.5rem;">
         <div class="header typo-header-3">
           <span class="id" on:click={navigateToUntracked}>{id}</span>
-          <FollowButton on:follow={follow} style="margin-left: 1rem;" />
+          <FollowToggle on:follow={follow} style="margin-left: 1rem;" />
         </div>
 
         <p style="color: var(--color-foreground-level-6);">
